fix: ResultGraph 분리
