+++ frontend/src/views/dashboard/components/ResultGraph.vue
@@ -21,7 +21,14 @@
           </div>
         </div>
       </div>
-      <div class="h-72 md:h-96"><canvas ref="lineCanvas"></canvas></div>
+
+      <PriceLineChart
+        :labels="processedData.labels"
+        :thisPrices="processedData.thisPrices"
+        :lastPrices="processedData.lastPrices"
+        :maxIdx="processedData.maxIdx"
+        :minIdx="processedData.minIdx"
+      />
     </div>
 
     <div class="pt-10 border-t border-slate-50 space-y-5">
@@ -33,7 +40,7 @@
           </h3>
           <p class="text-[13px] text-slate-500 mt-1 font-medium">
             조회 기간 전체 평균(<span class="font-mono font-bold text-slate-700 bg-slate-100 px-1 rounded"
-              >{{ Math.floor(avgPrice).toLocaleString() }}원</span
+              >{{ Math.floor(processedData.avgPrice).toLocaleString() }}원</span
             >) 대비 현황입니다.
           </p>
         </div>
@@ -42,14 +49,16 @@
           <span class="text-blue-400">하락 (-)</span>
         </div>
       </div>
-      <div class="h-40 md:h-52"><canvas ref="barCanvas"></canvas></div>
+
+      <PriceBarChart :labels="processedData.labels" :diffData="processedData.diffData" />
     </div>
   </div>
 </template>
 
 <script setup>
-import { ref, onMounted, watch, computed } from 'vue';
-import Chart from 'chart.js/auto';
+import { computed } from 'vue';
+import PriceLineChart from './PriceLineChart.vue';
+import PriceBarChart from './PriceBarChart.vue';
 
 const props = defineProps({
   chartData: {
@@ -59,27 +68,14 @@ const props = defineProps({
   },
 });
 
-const lineCanvas = ref(null);
-const barCanvas = ref(null);
-let lineChart = null;
-let barChart = null;
-
-const avgPrice = computed(() => {
-  const validPrices = props.chartData.thisPrices.filter((p) => p !== null);
-  return validPrices.length ? validPrices.reduce((a, b) => a + b, 0) / validPrices.length : 0;
-});
-
-const initCharts = () => {
-  if (!lineCanvas.value || !barCanvas.value) return;
-  if (lineChart) lineChart.destroy();
-  if (barChart) barChart.destroy();
-
+// 모든 계산 로직을 부모에서 통합 관리
+const processedData = computed(() => {
   const src = props.chartData;
-
   let labels = src.labels ? [...src.labels] : [];
   let thisPrices = src.thisPrices ? [...src.thisPrices] : [];
   let lastPrices = src.lastPrices ? [...src.lastPrices] : [];
 
+  // 데이터 1개일 때 처리
   if (labels.length === 1) {
     labels = ['', labels[0], ''];
     const mainThis = thisPrices[0] ?? null;
@@ -88,129 +84,34 @@ const initCharts = () => {
     lastPrices = [null, typeof mainLast === 'number' ? mainLast : null, null];
   }
 
-  const currentAvg = avgPrice.value;
-
+  // 평균 계산
   const validThisPrices = thisPrices.filter((p) => p !== null);
+  const avgPrice = validThisPrices.length ? validThisPrices.reduce((a, b) => a + b, 0) / validThisPrices.length : 0;
+
+  // 최대/최소값 인덱스 추출 (Point 점 표기용)
   const maxVal = Math.max(...validThisPrices);
   const minVal = Math.min(...validThisPrices);
   const maxIdx = thisPrices.indexOf(maxVal);
   const minIdx = thisPrices.indexOf(minVal);
 
+  // 등락 데이터 계산 (Bar Chart 전용)
   const diffData = thisPrices.map((val) => {
     if (val === null) return { diff: 0, pct: 0, val: 0 };
-    return { diff: val - currentAvg, pct: ((val - currentAvg) / currentAvg) * 100, val };
+    return {
+      diff: val - avgPrice,
+      pct: avgPrice !== 0 ? ((val - avgPrice) / avgPrice) * 100 : 0,
+      val,
+    };
   });
 
-  lineChart = new Chart(lineCanvas.value.getContext('2d'), {
-    type: 'line',
-    data: {
-      labels,
-      datasets: [
-        {
-          label: '금년',
-          data: thisPrices,
-          borderColor: '#475569',
-          borderWidth: 2,
-          pointRadius: thisPrices.map((_, i) => (i === maxIdx || i === minIdx ? 4 : 0)),
-          pointBackgroundColor: thisPrices.map((_, i) =>
-            i === maxIdx ? '#f43f5e' : i === minIdx ? '#3b82f6' : '#475569'
-          ),
-          pointBorderColor: '#fff',
-          tension: 0,
-          fill: false,
-          zIndex: 10,
-        },
-        {
-          label: '전년',
-          data: lastPrices,
-          borderColor: 'transparent',
-          backgroundColor: 'rgba(226, 232, 240, 0.6)',
-          pointRadius: 0,
-          tension: 0,
-          fill: true,
-          zIndex: 1,
-        },
-      ],
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      interaction: { mode: 'index', intersect: false },
-      plugins: { legend: { display: false } },
-      scales: {
-        x: { grid: { display: false }, ticks: { font: { size: 9 }, color: '#94a3b8' } },
-        y: {
-          grid: { color: '#f8fafc' },
-          ticks: { font: { size: 10, family: 'monospace' }, color: '#cbd5e1', callback: (v) => v.toLocaleString() },
-        },
-      },
-    },
-  });
-
-  barChart = new Chart(barCanvas.value.getContext('2d'), {
-    type: 'bar',
-    data: {
-      labels,
-      datasets: [
-        {
-          data: diffData.map((d) => d.diff),
-          backgroundColor: diffData.map((d) => (d.diff >= 0 ? '#fda4af' : '#93c5fd')),
-          hoverBackgroundColor: diffData.map((d) => (d.diff >= 0 ? '#f43f5e' : '#3b82f6')),
-          borderRadius: 2,
-          barPercentage: 0.8,
-        },
-      ],
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: { display: false },
-        tooltip: {
-          backgroundColor: 'rgba(30, 41, 59, 0.9)',
-          titleColor: '#fff',
-          bodyColor: '#fff',
-          bodyFont: { size: 11, weight: '500' },
-          padding: 12,
-          displayColors: false,
-          callbacks: {
-            label: (ctx) => {
-              const item = diffData[ctx.dataIndex];
-              const prefix = item.diff >= 0 ? '▲' : '▼';
-              const sign = item.diff >= 0 ? '+' : '';
-
-              return [
-                ` 실거래가: ${item.val.toLocaleString()}원`,
-                ` 평균대비: ${prefix} ${Math.abs(Math.floor(item.diff)).toLocaleString()}원 (${sign}${item.pct.toFixed(
-                  1
-                )}%)`,
-              ];
-            },
-          },
-        },
-      },
-      scales: {
-        x: { display: false },
-        y: {
-          suggestedMin: Math.min(...diffData.map((d) => d.diff)) * 1.5,
-          suggestedMax: Math.max(...diffData.map((d) => d.diff)) * 1.5,
-          grid: {
-            color: (ctx) => (ctx.tick.value === 0 ? '#94a3b8' : 'transparent'),
-            lineWidth: 2,
-          },
-          ticks: {
-            display: true,
-            values: [0],
-            callback: (val) => (val === 0 ? 'AVG' : ''),
-            font: { size: 9, weight: 'bold' },
-            color: '#94a3b8',
-          },
-        },
-      },
-    },
-  });
-};
-
-onMounted(initCharts);
-watch(() => props.chartData, initCharts, { deep: true });
+  return {
+    labels,
+    thisPrices,
+    lastPrices,
+    avgPrice,
+    maxIdx,
+    minIdx,
+    diffData,
+  };
+});
 </script>
